+++ springboot/query-core/src/test/java/org/fuin/cqrs4j/springboot/query/core/ArchitectureTest.java
@@ -30,6 +30,7 @@ public class ArchitectureTest {
             .should()
             .onlyDependOnClassesThat()
             .resideInAnyPackage(THIS_PACKAGE, "java..",
+                    "io.micrometer..",
                     "javax.sql..",
                     "jakarta.persistence..",
                     "jakarta.validation..",